requirements.txt
openai==0.27.2
tweepy==4.14.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.5.0
textblob==0.17.1
nltk==3.8.1
numpy==1.25.2
schedule==1.2.0
runwayml
rapidfuzz
pillow==11.1.0
